ROSE/requirements.txt
Pillow
einops
safetensors
timm
tomesd
torch>=2.1.2
torchdiffeq
torchsde
decord
datasets
numpy
scikit-image
opencv-python
omegaconf
SentencePiece
albumentations
imageio[ffmpeg]
imageio[pyav]
tensorboard
beautifulsoup4
ftfy
func_timeout
accelerate>=0.25.0
diffusers>=0.30.1,<=0.31.0
# transformers 5.x removes symbols diffusers<=0.31 still imports (e.g. FLAX_WEIGHTS_NAME)
transformers>=4.46.2,<5.0
